Simplify macro invocation
The macro used a 'clever' construction with a single basecase, coupled with
multiple refinements that would call eachother.

Should the user input an unexpected token (like a trailing comma) this would
cause those refinement branches to no longer apply, and the base case getting
called again.

Since it is a catch-all tt* branch, it would recursively call into the
refinement again and again. The problem was that this expanded to _twice_ the
amount of code. So it caused an exponential slow-down in the compiler until it
did not progress further. The fix was to... just no longer do that as it
was not needed anymore.

Signed-off-by: Marcel Müller <neikos@neikos.email>